Update method of grid initialisation
MY intial method of implementing the grid was fine, however the way I
build the grid using the user input is different though more efficient.
Therefore, for consistency, I have updated the logic to build the
initial grid.
However, the grid is not currently displaying as a result of another
layer of elements