Fix restoring element bug
After grid size is altered, after a grid unit was hovered and returned
to its original state its size was not retained from the size it was
when the grid was built.
Implementing a page counter and getting the grid length value using
reg ex matching was used to solve this issue.